Wind prediction example for scipy.interpolate
This revision starts the euroscipy tutorial 2010 on scipy by adding a
wind speed prediction example illustrating the UnivariateSpline use
from the scipy.interpolate module. As a result, the scipy file:
    source/scipy.rst
has been added and is currently the only one of the index:
    source/index.rst

In the new data directory, the script generating the result
figure, cumulative-wind-speed-prediction.png, has been added:
    data/cumulative-wind-speed-prediction.py
and it uses the annual maxima wind speeds stored in the file:
    data/max-speeds.npy

For interested readers, an exercice is proposed based on the
raw measurements data found in:
    data/sprog-windspeeds.npy
The correction is provided by the script:
    data/gumbell-wind-speed-prediction.py
which generate two figures containing results.

For ensuring that scripts work correclty and saving space
on the repo, two new rules have been added to:
    Makefile
where figures are generated before building html documentation.
From now, the two working commands are:
    make test
    make html