Fix the interpolation code to match the images
The code examples for interpolating the face image show no option being passed to `interpolation`, then `nearest` being passed. 
However the images first show a smooth interpolation, followed by the rough `nearest`. I'm not sure if the default used to be bilinear, but now when you run the examples, the default of `interpolation=none` is the same as `nearest`, so you get two of the same result if you copy paste the code.

I also added the link to the matplotlib interpolation example because it is a quick illustrator to the types you can choose from