ENH: Make the numpy advanced iterator example more elaborate
This uses the iterator to allocate the output. It also adds some
checks so that the code will not fail on non-native byte order
or unalinged inputs.
It also makes use of the inner loop optimization which should be
used for most iterations such as here.

Possibly this goes too far and a simpler version is better?

Closes gh-402.